# [#109940854] As a developer, I would like the server-side template engine switched to JTS so that the platform can experience better performance during peak traffic periods.