--- a/config/views.js
+++ b/config/views.js
@@ -10,6 +10,11 @@
  * For more information on views and layouts, check out:
  * http://sailsjs.org/#!/documentation/concepts/Views
  */
+var JTS = require('jts');
+var engine = new JTS({
+  defaultLayout: 'layout',
+  layouts: 'views'
+});
 
 module.exports.views = {
 
@@ -30,66 +35,9 @@ module.exports.views = {
   *                                                                           *
   ****************************************************************************/
 
-  engine: 'ejs',
-
-
-  /****************************************************************************
-  *                                                                           *
-  * Layouts are simply top-level HTML templates you can use as wrappers for   *
-  * your server-side views. If you're using ejs or jade, you can take         *
-  * advantage of Sails' built-in `layout` support.                            *
-  *                                                                           *
-  * When using a layout, when one of your views is served, it is injected     *
-  * into the `body` partial defined in the layout. This lets you reuse header *
-  * and footer logic between views.                                           *
-  *                                                                           *
-  * NOTE: Layout support is only implemented for the `ejs` view engine!       *
-  *       For most other engines, it is not necessary, since they implement   *
-  *       partials/layouts themselves. In those cases, this config will be    *
-  *       silently ignored.                                                   *
-  *                                                                           *
-  * The `layout` setting may be set to one of the following:                  *
-  *                                                                           *
-  * If `false`, layouts will be disabled. Otherwise, if a string is           *
-  * specified, it will be interpreted as the relative path to your layout     *
-  * file from `views/` folder. (the file extension, ".ejs", should be         *
-  * omitted)                                                                  *
-  *                                                                           *
-  ****************************************************************************/
-
-  /****************************************************************************
-  *                                                                           *
-  * Using Multiple Layouts                                                    *
-  *                                                                           *
-  * If you're using the default `ejs` or `handlebars` Sails supports the use  *
-  * of multiple `layout` files. To take advantage of this, before rendering a *
-  * view, override the `layout` local in your controller by setting           *
-  * `res.locals.layout`. (this is handy if you parts of your app's UI look    *
-  * completely different from each other)                                     *
-  *                                                                           *
-  * e.g. your default might be                                                *
-  * layout: 'layouts/public'                                                  *
-  *                                                                           *
-  * But you might override that in some of your controllers with:             *
-  * layout: 'layouts/internal'                                                *
-  *                                                                           *
-  ****************************************************************************/
-
-  layout: 'layout',
-
-  /****************************************************************************
-  *                                                                           *
-  * Partials are simply top-level snippets you can leverage to reuse template *
-  * for your server-side views. If you're using handlebars, you can take      *
-  * advantage of Sails' built-in `partials` support.                          *
-  *                                                                           *
-  * If `false` or empty partials will be located in the same folder as views. *
-  * Otherwise, if a string is specified, it will be interpreted as the        *
-  * relative path to your partial files from `views/` folder.                 *
-  *                                                                           *
-  ****************************************************************************/
-
-  partials: false
-
+  engine: {
+    ext: 'jts',
+    fn: engine.render,
+  }
 
 };