--- a/views/layout.jts
+++ b/views/layout.jts
@@ -45,7 +45,7 @@
     <!--ANGULAR VIEW-->
   </div>
 
-  <%- body %>
+  ${ body }
 
   <script src="/socket.io/socket.io.js"></script>
   <script src="https://ajax.googleapis.com/ajax/libs/angularjs/1.4.4/angular.min.js"></script>