--- a/views/podcast/embed.jts
+++ b/views/podcast/embed.jts
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html>
+<head>
+  <title>${ episode.name } powered by Bethel</title>
+  <link rel="stylesheet" href="/components/videojs/dist/video-js.min.css" />
+  <link rel="stylesheet" href="/styles/embed.css" />
+  <style>${ embedStyles }</style>
+</head>
+<body>
+
+  <${ podcast.type } id="embedded" class="video-js vjs-bethel-skin" uuid="${ episode.id }" ${ (podcast.embedSettings && podcast.embedSettings.feedUrl) ? `feed="${ podcast.embedSettings.feedUrl }"` : '' } controls poster="${ posterImage }" preload="none">
+
+  ${ !episode.variants ? `
+    <source src="/podcastmedia/download/${ episode.id }?embed" type="${ PodcastHelper.mimeTypeFromUrl(episode.url) }" />
+  ` : `
+    ${ episode.variants.hls ? `<source src="/podcastmedia/download/${ episode.id }?variant=hls&embed" type="application/x-mpegURL" />` : '' }
+    <source src="/podcastmedia/download/${ episode.id }?variant=mp4&embed" type="video/mp4" />
+  ` }
+
+    <p class="vjs-no-js">Please enable JavaScript, and consider upgrading to a web browser that <a href="http://videojs.com/html5-video-support/" target="_blank">supports HTML5 video</a></p>
+  </${ podcast.type }>
+
+  ${ podcast.type === 'audio' ? `<div class="meta">
+    <h1>${ episode.name }</h1>
+    <h2>${ podcast.name }${ episode.date ? `<br /><small>${ episode.date.toLocaleDateString() }</small>` : '' }</h2>
+  </div>` : '' }
+
+  <script src="/components/videojs/dist/video.min.js"></script>
+  <script src="/min/embed.min.js"></script>
+</body>
+</html>