--- a/views/rss.jts
+++ b/views/rss.jts
@@ -1,5 +1,5 @@
 <rss xmlns:itunes="http://www.itunes.com/dtds/podcast-1.0.dtd" xmlns:atom="http://www.w3.org/2005/Atom" version="2.0">
 <channel>
-<%- body %>
+${ body }
 </channel>
 </rss>