- Migrates all server templates from EJS to JTS
- Encodes HTML entities on podcast feed XML to prevent parsing errors

I ran a quick and dirty benchmark locally to compare the performance gains. For this test I used the Harvest podcast since it has a large number of episodes. This isn't likely 100% accurate since I'm load testing from the machine the server is running locally. The following command was used: 

`loadtest -t 60 -c 10 --rps 100 http://localhost:1337/podcast/feed/54de3b14deb7d7090079dfc4`

**EJS**
```
Completed requests:  4239
Total errors:        43
Total time:          60.004157352 s
Requests per second: 71
Total time:          60.004157352 s

Percentage of the requests served within a certain time
 50%      5419 ms
 90%      12712 ms
 95%      13788 ms
 99%      16798 ms
100%      21019 ms (longest request)
```

**JTS**
```
Completed requests:  5331
Total errors:        0
Total time:          60.002507531000006 s
Requests per second: 89
Total time:          60.002507531000006 s

Percentage of the requests served within a certain time
 50%      937 ms
 90%      3913 ms
 95%      5401 ms
 99%      6299 ms
100%      6984 ms (longest request)
```

Based on these results JTS is seeing an average load speed increase of **5.7X** in the 50th percentile and **3.2X** in the 90th percentile at a sustained load of 100/req/s.