@Ignigena 2015-12-11
@Josh-Mosh E2E tests are passing but I'd love a second set of eyes especially where these tests don't cover. In particular the invite screen and the podcast embed and feed endpoints. Chrome will complain of any errors in XML parsing on the feed if there are unescaped HTML entities.