@Ignigena 2016-01-29
Update on the benchmarks after switching off Keen and moving to Elasticsearch:

**EJS**
```
Completed requests:  5760
Total errors:        0
Total time:          60.004210164 s
Requests per second: 96
Total time:          60.004210164 s

Percentage of the requests served within a certain time
  50%      52 ms
  90%      816 ms
  95%      1105 ms
  99%      1786 ms
 100%      2106 ms (longest request)
```

**JTS**
```
Completed requests:  5929
Total errors:        0
Total time:          60.004246908 s
Requests per second: 99
Total time:          60.004246908 s

Percentage of the requests served within a certain time
  50%      26 ms
  90%      498 ms
  95%      847 ms
  99%      1396 ms
 100%      1557 ms (longest request)
```

Based on these benchmarks we're seeing speed improvements anywhere from **1.3X** on the 95th percentile to **2.0X** on the 50th percentile.